refactored icoremodel to external model
