+++ src/main/kotlin/org/ionproject/integration/model/external/generic/CoreAcademicCalendar.kt
@@ -1,3 +1,3 @@
 package org.ionproject.integration.model.external.generic
 
-class CoreExamSchedule() : ICoreModel
+class CoreAcademicCalendar() : ICoreModel
