+++ src/main/kotlin/org/ionproject/integration/model/external/generic/ICoreModel.kt
@@ -1,3 +1,3 @@
 package org.ionproject.integration.model.external.generic
 
-class CoreExamSchedule() : ICoreModel
+interface ICoreModel
