+++ src/main/kotlin/org/ionproject/integration/model/internal/generic/AcademicCalendar.kt
@@ -1,5 +1,7 @@
 package org.ionproject.integration.model.internal.generic
 
+import org.ionproject.integration.model.external.generic.CoreAcademicCalendar
+import org.ionproject.integration.model.external.generic.ICoreModel
 import org.ionproject.integration.model.internal.timetable.School
 
 data class AcademicCalendar(
