+++ src/main/kotlin/org/ionproject/integration/model/internal/generic/ExamSchedule.kt
@@ -1,6 +1,7 @@
 package org.ionproject.integration.model.internal.generic
 
 import org.ionproject.integration.model.external.generic.CoreExamSchedule
+import org.ionproject.integration.model.external.generic.ICoreModel
 import org.ionproject.integration.model.internal.timetable.School
 
 data class ExamSchedule(
