--- src/main/kotlin/org/ionproject/integration/model/internal/generic/ICoreModel.kt
@@ -1,3 +1,0 @@
-package org.ionproject.integration.model.internal.generic
-
-interface ICoreModel
