+++ src/main/kotlin/org/ionproject/integration/model/internal/generic/IInternalModel.kt
@@ -1,5 +1,7 @@
 package org.ionproject.integration.model.internal.generic
 
+import org.ionproject.integration.model.external.generic.ICoreModel
+
 interface IInternalModel {
     fun toCore(): ICoreModel
 }
