+++ src/main/kotlin/org/ionproject/integration/model/internal/generic/Timetable.kt
@@ -1,5 +1,7 @@
 package org.ionproject.integration.model.internal.generic
 
+import org.ionproject.integration.model.external.generic.ICoreModel
+
 class Timetable : IInternalModel {
     override fun toCore(): ICoreModel {
         TODO("Not yet implemented")
