+++ src/main/kotlin/org/ionproject/integration/step/chunkbased/processor/GenericProcessor.kt
@@ -1,6 +1,6 @@
 package org.ionproject.integration.step.chunkbased.processor
 
-import org.ionproject.integration.model.internal.generic.ICoreModel
+import org.ionproject.integration.model.external.generic.ICoreModel
 import org.ionproject.integration.model.internal.generic.IInternalModel
 import org.springframework.batch.core.configuration.annotation.StepScope
 import org.springframework.batch.item.ItemProcessor
