+++ src/main/kotlin/org/ionproject/integration/step/chunkbased/writer/GenericCoreWriter.kt
@@ -1,6 +1,6 @@
 package org.ionproject.integration.step.chunkbased.writer
 
-import org.ionproject.integration.model.internal.generic.ICoreModel
+import org.ionproject.integration.model.external.generic.ICoreModel
 import org.springframework.batch.core.configuration.annotation.StepScope
 import org.springframework.batch.item.ItemWriter
 import org.springframework.stereotype.Component
